+++ CHANGELOG.md
@@ -37,3 +37,4 @@ See [keep a changelog](https://keepachangelog.com/en/1.0.0/) for information abo
 * Fixed budget path.
 * Fixed planning js.
 * Fixed filtering issue with project overview.
+* Changed sprint report form to GET method.